fix(*): In GitHub Markdown, SVG elements are block elements.
Since it is a block element, there is a problem that it is automatically
opened. To compensate for this, it is changed to the following form.

- AS-IS: svg [name](url)
- TO-BE: [svg name](url)

Code Block: 'Plain text' is not supported (a supported language format
when specifying a code block in GitHub Markdown)

Does not work;

```plain text
```

TO BE;

```
```