fix(*): The configuration of the original url and the content is the same.
The use configuration of the content is the same as the original content
configuration. Previously, it was implemented by organizing markdown
codes and code block areas, but in some cases, even though the code
was written correctly, the same result screen as the original content
could not be derived. Accordingly, the previous markdown code area
was replaced with an image.